extract some TEE-specific change grouping logic + refactor
