+++ src/patchwork/components/MinimapWithDiff.tsx
@@ -1,7 +1,7 @@
 import { MarkdownDoc } from "@/tee/schema";
 import { Patch } from "@automerge/automerge/next";
 import { truncate } from "lodash";
-import { Heading, extractHeadings } from "../groupChanges";
+import { Heading, extractHeadings } from "@/tee/statsForChangeGroup";
 import { TextPatch } from "../utils";
 
 function patchOverlapsLine(start: number, end: number, patch: Patch): boolean {
